moving beforeAll processes around
