+++ webapp/src/test/integration.test.ts
@@ -2,3 +2,5 @@ import './account';
 import './pages';
 import './teams';
 import './models';
+
+import './logout'; //this MUST be last
